The following new subprograms have been added. 	initpep         : Preperation of extended initial structures for folding simulation. 	pdb_movie       : Generation of a multi-model PDB file from sequential single-structure PDB files for trajectory visualization.
Changed the default electrostatics method in “md_run” from Zero-Dipole summation Method (ZD) to Cell Multipole Method (CMM).
Slightly modified the behavior of the INPENE option in “batch_set”.
Fixed a bug in prep that occurred when an END line was present at the end of the input PDB file.
Adjusted log output related to the “atom_spec_PCA” option in “Ens_Ana”.
Improved the calculation method of ALSD simulation parameters in “GEprep” for greater stability.